+++ terraform/modules/aws-eks-cluster/outputs.tf
@@ -1,0 +1,28 @@
+output "cluster_name" {
+  value = aws_eks_cluster.eks_cluster.name
+}
+
+output "cluster_endpoint" {
+  value = aws_eks_cluster.eks_cluster.endpoint
+}
+
+output "cluster_ca_certificate" {
+  value     = aws_eks_cluster.eks_cluster.certificate_authority[0].data
+  sensitive = true
+}
+
+output "cluster_security_group_id" {
+  value = aws_security_group.eks_cluster_sg.id
+}
+
+output "oidc_provider_arn" {
+  value = aws_iam_openid_connect_provider.eks_oidc.arn
+}
+
+output "oidc_provider_url" {
+  value = replace(aws_eks_cluster.eks_cluster.identity[0].oidc[0].issuer, "https://", "")
+}
+
+output "cluster_iam_role_arn" {
+  value = aws_iam_role.eks_cluster_role.arn
+}