+++ terraform/modules/aws-eks-vpc/outputs.tf
@@ -1,0 +1,15 @@
+output "vpc_id" {
+  value = aws_vpc.eks_vpc.id
+}
+
+output "public_subnet_ids" {
+  value = aws_subnet.public[*].id
+}
+
+output "private_subnet_ids" {
+  value = aws_subnet.private[*].id
+}
+
+output "nat_gateway_ids" {
+  value = aws_nat_gateway.eks_nat[*].id
+}